Fix DNS resolvers removal on disconnect
By default openconnect runs as `nobody`, and the disconnect part
of the script is run under that user. However, it is very unlikely
that in default MacOS setups that user is allowed to run scutil,
nor it would be a good practice to do so.

So, the README now suggests to run openconnect as the calling user,
and `sudo` are added to the connect/disconnect scripts.